Basic authentication clarification
We no longer use the private OData API, so it's not necessary to
preemptively send the Authorization header. But since we know that
basic authentication will always be required, it's fine to keep
sending the header rather than using the Java 11 HTTP client
Authenticator. This means one less round-trip to the server.