Added the -skip-drafts command-line option
You can now use the -skip-drafts command-line option to ignore iflows
that are in the draft state. The change is implemented in the Xquery
file iflow-artifact-ids-from-api-response.xquery and in the
CloudIntegrationOdataApi class. The getIflowArtifactIdsFromPackage
method in the CloudIntegrationApi interface was updated to accept a
boolean parameter, indicating whether or not to skip draft iflows.

The CliClient class was updated to understand the new command-line
option and to pass it on to the TenantAllArtifactsSupplier class, which
calls the API.