Added the -packages command-line option
In addition to inspecting individual iflows and all iflows in the
tenant, you can now inspect the iflows in one or more packages, with the
option of skipping drafts and skipping certain iflows.

TenantIndividualPackagesSupplier is the new IflowArtifactSupplier class.
Since there was quite a lot of overlap between the setup() methods in
TenantAllArtifactsSupplier and the new supplier, the shared code has
been moved to a new superclass (PackageSupplierBase).