Xerces added for XML Schema 1.1 validation
To be able to do XML Schema 1.1 validation, Apache Xerces has been
added to the project. Please note that the version of Xerces in the
Maven repository doesn't support XML Schema 1.1. Therefore, the required
JAR files have been added to the CPILint repository and are copied into
the distribution archive.

If/when the version of Xerces in the Maven repository supports XML
Schema 1.1, we can just add it as an Ivy dependency like all the other
libraries we use.

With XML Schema 1.1, the schema now validates all aspects of the rules
file. This makes some code in RulesFile.validateRules redundant, and
this code has been removed.