CleartextBasicAuthNotAllowed update
The CleartextBasicAuthNotAllowed rule was updated with support for the
new Anaplan and Jira receiver adapters.

Please note that the endpoints in question are API and authentication
endpoints for the respective services, so channels likely cannot be
configured correctly with unencrypted endpoints. Nevertheless, the rule
now checks channels using these new receiver adapters as well.