Display package information in CPILint output
In addition to displaying the name and ID of an iflow that is not in
compliance, CPILint will now also show the name and ID of that iflow's
package.

The additional package information is displayed when you check the
entire tenant or individual packages. It is not displayed when checking
individual iflows in the tenant or when checking iflow artifact files on
the local machine.

Please note that from this commit, the required Java version is 17. Java
17 is the next Long Term Support (LTS) release after Java 11.

Closes #18