New rule exemption feature
CPILint now supports declaring that certain iflows are exempt from
certain rules. You do this in the new <exemptions> element in the
rules file.

In order to reference a specific rule in an exemption, rules now have an
optional ID and issues now contain the ID of the rule that created the
issue. This is 100% backward compatible with existing rules files.

The change to issues impacts all existing implementations of issues and
rules, which is why this particular change touched rather a lot of
files.