Added the unpacked files supplier
CPILint now supports checking iflows that are stored unpacked locally.
I.e. not as a ZIP archive, as when you download an iflow, but
uncompressed as when iflows are stored in version control.

Running CPILint in CI/CD pipelines is the main use case for this new
feature. The new command-line option is called -unpacked-files.